+++ packages/dids/src/did.ts
@@ -5,7 +5,8 @@ import { RPCClient } from 'rpc-utils'
 import { CID } from 'multiformats/cid'
 import { CacaoBlock, Cacao, Verifiers } from '@didtools/cacao'
 import { getEIP191Verifier } from '@didtools/pkh-ethereum'
-import type { DagJWS, DIDProvider, DIDProviderClient } from './types.js'
+import type { DagJWS } from '@didtools/codecs'
+import type { DIDProvider, DIDProviderClient } from './types.js'
 import {
   fromDagJWS,
   encodeBase64,